refactor: decompose database-view-client.tsx into focused hooks and sub-components (#670)
